pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-deepgram>=1.2.9",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.2.9",
    "livekit-plugins-silero>=1.2.9",
    "livekit-plugins-turn-detector>=1.2.9",
    "huggingface_hub>=0.23,<1.0",
    "python-dotenv",
    "llama-index-core>=0.13.0,<0.14",
    "llama-index-llms-openai>=0.5.0,<0.6",
    "llama-index-postprocessor-flag-embedding-reranker>=0.1.2,<1.0",
    "FlagEmbedding>=1.2.0,<2.0",
    "torch>=2.2,<3.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "fakeredis",
    "pytest-mock",
    "pytest-cov",
    "respx",
    "fastapi",
    "ruff",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "fakeredis",
    "pytest-mock",
    "pytest-cov",
    "respx",
    "fastapi",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
